requirement.txt
# Linux dependencies
evdev; sys_platform == "linux"

# macOS dependencies
hid; sys_platform == "darwin"

dotenv
paho-mqtt
python-escpos
pyserial
